==> Part 2A - Data Extraction from TMDB API.ipynb <==
# Data Extraction from TMDB API

# Import Libraries 
import pandas as pd
import numpy as np
import os, json, time
from tqdm.notebook import tqdm_notebook 

## User Defined Functions

### get_movie_with_rating Function

# function to get movie raiting from tmdb api
def get_movie_with_rating(movie_id):
    """Adapted from source = https://github.com/celiao/tmdbsimple"""
    # get movie object for current id
    movie = tmdb.Movies(movie_id)
    
    # save the .info .releases dictionaries
    info = movie.info()
    releases = movie.releases()
    
    # loop through countries in releases
    for c in releases['countries']:
        # if country abbreviation==US
        if c['iso_3166_1'] == 'US':
            # save a 'certification' key in info with the certification
            info['certification'] = c['certification']
    
    return info

#### Test get_movie_with_rating Function

Test with The Avengers

test1 = get_movie_with_rating('tt0848228')
test1

Test with The Notebook 

test2 = get_movie_with_rating('tt0332280')
test2

The function works as predicted.

### write_json Function

# appends a list of records (new_data) to a json file
# addapted from : https://www.geeksforgeeks.org/append-to-json-file-using-python/

def write_json(new_data, filename):
    with open(filename,'r+') as file:
        # first load existing data into dict
        file_data = json.load(file)
        # choose extend or append
        if (type(new_data) == list) & (type(file_data) == list):
            file_data.extend(new_data)
        else:
            file_data.append(new_data)
        # set file's current position at offset
        file.seek(0)
        # convert back to json
        json.dump(file_data, file)

#### Test write_json Function

# create new json file with empty list
JSON_FILE = 'Data/test_json.json'
with open(JSON_FILE,'w') as f:
    json.dump([],f)
write_json(test1, JSON_FILE)
write_json(test2, JSON_FILE)

# save json file as dataframe and look at info
test_df = pd.read_json(JSON_FILE)
test_df

The function works as predicted. 

## Extract Data from TMDB

### tmdbsimple Package

!pip install tmdbsimple;
import tmdbsimple as tmdb

# load login credentials using json file in notebook
with open('E:/Dropbox/Coding Dojo/Projects/Project 3/movie_success_predictions/.secret/tmdb_api.json', 'r') as f:
    login = json.load(f)
# display keys of loaded dict
login.keys()

# save api-key using tmdbsimple
tmdb.API_KEY = login['api-key']

### Extract data from TMDB API

# define list of years to extract
YEARS_TO_GET = [2000, 2001]
# define empty errors list
errors = []

# set up progress bar as start of outer loop
for YEAR in tqdm_notebook(YEARS_TO_GET, desc = 'YEARS', position=0):
    
    # select JSON_FILE name and check if file exists
    JSON_FILE = f'Data/tmdb_api_results_{YEAR}.json'
    # check if JSON_FILE already exists or not
    file_exists = os.path.isfile(JSON_FILE)
    
    # if file does not exist, create it
    if file_exists == False:
        # save empty dict with just 'imdb_id' to the new json file
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)
            
    # save new year as the current df
    df = basics.loc[basics['startYear']==YEAR].copy()
    # save movie ids to list
    movie_ids = df['tconst'].copy()
    
    # load existing data from json to dataframe
    previous_df = pd.read_json(JSON_FILE)
    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]
    
    # get index and movie id from list
    # INNER LOOP
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                 desc=f'Movies from {YEAR}',
                                 position=1,
                                 leave=True):
        try:
            # retrieve then data for the movie id
            temp = get_movie_with_rating(movie_id)
            # append/extend results to existing file using user defined functions
            write_json(temp, JSON_FILE)
            # short 20ms sleep to prevent overwhelming server
            time.sleep(.02)
            
        except Exception as e:
            # append errors to errors list
            errors.append([movie_id,e])
            
    # convert json file to .csv.gz
    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f'Data/final_tmdb_data_{YEAR}.csv.gz', 
                         compression='gzip',
                        index=False)
    
print(f'- Total errors: {len(errors)}')

## Extract data from 2010 - 2020

YEARS_TO_GET = list(range(2010, 2021))
# define empty errors list
errors = []

basics = pd.read_csv('Data/title_basics.csv.gz')

# set up progress bar as start of outer loop
for YEAR in tqdm_notebook(YEARS_TO_GET, desc = 'YEARS', position=0):
    
    # select JSON_FILE name and check if file exists
    JSON_FILE = f'Data/tmdb_api_results_{YEAR}.json'
    # check if JSON_FILE already exists or not
    file_exists = os.path.isfile(JSON_FILE)
    
    # if file does not exist, create it
    if file_exists == False:
        # save empty dict with just 'imdb_id' to the new json file
        with open(JSON_FILE,'w') as f:
            json.dump([{'imdb_id':0}],f)
            
    # save new year as the current df
    df = basics.loc[basics['startYear']==YEAR].copy()
    # save movie ids to list
    movie_ids = df['tconst'].copy()
    
    # load existing data from json to dataframe
    previous_df = pd.read_json(JSON_FILE)
    # filter out any ids that are already in the JSON_FILE
    movie_ids_to_get = movie_ids[~movie_ids.isin(previous_df['imdb_id'])]
    
    # get index and movie id from list
    # INNER LOOP
    for movie_id in tqdm_notebook(movie_ids_to_get,
                                 desc=f'Movies from {YEAR}',
                                 position=1,
                                 leave=True):
        try:
            # retrieve then data for the movie id
            temp = get_movie_with_rating(movie_id)
            # append/extend results to existing file using user defined functions
            write_json(temp, JSON_FILE)
            # short 20ms sleep to prevent overwhelming server
            time.sleep(.02)
            
        except Exception as e:
            # append errors to errors list
            errors.append([movie_id,e])
            
    # convert json file to .csv.gz
    final_year_df = pd.read_json(JSON_FILE)
    final_year_df.to_csv(f'Data/final_tmdb_data_{YEAR}.csv.gz', 
                         compression='gzip',
                        index=False)
    
print(f'- Total errors: {len(errors)}')
