==> Part 3 - Create MySQL Database from Movie Data.ipynb <==
# Creating a MySQL Database from Movie Data

# Import Libraries 
import pandas as pd
import numpy as np

import pymysql
from sqlalchemy import create_engine
from sqlalchemy.types import *
from sqlalchemy_utils import database_exists, create_database

# read in dataframes from GitHub Data folder
title_basics = pd.read_csv('Data/title_basics.csv.gz')
title_ratings = pd.read_csv('Data/title_ratings.csv.gz')
tmdb_data = pd.read_csv('Data/tmdb_results_combined.csv.gz')

## Load and Explore Dataframes

### Title Basics

title_basics.info()
title_basics.head()

# for the title basics, we only need the tconst, primary title, start year,
# runtime, and genres. I will resave the df with only these columns
title_basics.drop(columns=['titleType','originalTitle','isAdult','endYear'],
                    inplace=True)
title_basics.info()
title_basics.head()

### Title Ratings 

title_ratings.info()
title_ratings.head()

### TMDB API Results

tmdb_data.info()
tmdb_data.head()

# for the api_df we only need the movie id, revenue, budget, and certification
# I will save the df only keeping these 4 columns
tmdb_data = tmdb_data[['imdb_id', 'revenue', 'budget', 'certification']]
tmdb_data.info()
tmdb_data.head()

I will delete all rows in which the imdb_id == 0, since these are not actual movie data.

tmdb_data = tmdb_data[tmdb_data['imdb_id'] != '0']
tmdb_data.head()

## Normalize Title Basics Table

### Normalize genres

Convert single string of genres from the title basics dataframe into two new tables: title_genres and genres

# create new column with a list of all unique genres for each title
# using the split function
title_basics['genres_split'] = title_basics['genres'].str.split(',')
title_basics.head(2)

# use .explode() to separate list of genres into new rows
exploded_genres = title_basics.explode('genres_split')

# get string column converted to a list of strings in each row
genres_split = title_basics['genres'].str.split(',')
# explode the series and take unique entries only
unique_genres = genres_split.explode().unique()
# sort unique genres
unique_genres = sorted(exploded_genres['genres_split'].unique())
unique_genres

# create new title_genres table
title_genres = exploded_genres[['tconst','genres_split']].copy()
title_genres.head()

# create genre mapper dictionary
genre_ints = range(len(unique_genres))
genre_map = dict(zip(unique_genres, genre_ints))
genre_map

# replace genres in title_genres with new integer id
title_genres['genre_id'] = title_genres['genres_split'].map(genre_map)
title_genres.drop(columns='genres_split', inplace=True)

# convert genre map dictionary to dataframe
genres = pd.DataFrame({'Genre_Name':genre_map.keys(),
                      'Genre_ID':genre_map.values()})
genres.head()

### Discard unnecessary information

# drop genres and genres_split column from title_basics dataframe
title_basics.drop(columns=['genres','genres_split'], inplace=True)
title_basics.head()

## Saving MySQL Tables

### Create MySQL Server

# Server information
username = "dcbailo89"
password = "1qaz2wsx3edc4rfv"
db_name = "movies"
host_name = "192.168.1.6"
connection = f"mysql+pymysql://{username}:{password}@{host_name}/{db_name}"

engine = create_engine(connection)
if not database_exists(engine.url):
    create_database(engine.url)
print(database_exists(engine.url))
engine

ABOVE CODE TO CREATE A NEW ENGINE FOUND ON STACK OVERFLOW: https://stackoverflow.com/questions/6506578/how-to-create-a-new-database-using-sqlalchemy

### title_basics Table

# view title_basics dataframe
title_basics.info()
title_basics.head()

# calculate max string lengths for object columns
key_len = title_basics['tconst'].fillna('').map(len).max()
title_len = title_basics['primaryTitle'].fillna('').map(len).max()
# create schema dictionary using SQLAlchemy datatype objects
title_basics_schema = {
    "tconst":String(key_len+1),
    "primaryTitle":Text(title_len+1),
    "startYear":Integer(),
    "runtimeMinutes":Integer()}

# add table to MySQL Server
title_basics.to_sql('title_basics',engine,dtype=title_basics_schema,
                   if_exists='replace',index=False)

# add primary key to title_basics table
engine.execute('ALTER TABLE title_basics ADD PRIMARY KEY (`tconst`);')

# execute query to display first 5 rows of table
q = """
SELECT * FROM title_basics
LIMIT 5;
"""
pd.read_sql(q,engine)

### title_ratings Table

# view title_ratings dataframe
title_ratings.info()
title_ratings.head()

# calculate max string lengths for object columns
key_len = title_ratings['tconst'].fillna('').map(len).max()
# create schema dictionary using SQLAlchemy datatype objects
title_ratings_schema = {
    "tconst":String(key_len+1),
    "averageRating":Float(),
    "numVotes":Integer()}

# add table to MySQL Server
title_ratings.to_sql('title_ratings',engine,dtype=title_ratings_schema,
                   if_exists='replace',index=False)

# add primary key to title_basics table
engine.execute('ALTER TABLE title_ratings ADD PRIMARY KEY (`tconst`);')

# execute query to display first 5 rows of table
q = """
SELECT * FROM title_ratings
LIMIT 5;
"""
pd.read_sql(q,engine)

### title_genres Table

# view title_genres dataframe
title_genres.info()
title_genres.head()

# calculate max string lengths for object columns
key_len = title_genres['tconst'].fillna('').map(len).max()
# create schema dictionary using SQLAlchemy datatype objects
title_genres_schema = {
    "tconst":String(key_len+1),
    "genre_id":Integer()}

# add table to MySQL Server
title_genres.to_sql('title_genres',engine,dtype=title_genres_schema,
                   if_exists='replace',index=False)

# execute query to display first 5 rows of table
q = """
SELECT * FROM title_genres
LIMIT 5;
"""
pd.read_sql(q,engine)

### genres Table

# view genres dataframe
genres.info()
genres.head()

# calculate max string lengths for object columns
name_len = genres['Genre_Name'].fillna('').map(len).max()
# create schema dictionary using SQLAlchemy datatype objects
genres_schema = {
    "Genre_Name":String(name_len+1),
    "Genre_ID":Integer()}

# add table to MySQL Server
genres.to_sql('genres',engine,dtype=genres_schema,
                   if_exists='replace',index=False)

# add primary key to title_basics table
engine.execute('ALTER TABLE genres ADD PRIMARY KEY (`Genre_ID`);')

# execute query to display first 5 rows of table
q = """
SELECT * FROM genres
LIMIT 5;
"""
pd.read_sql(q,engine)

### tmdb_data Table

# view tmdb_data dataframe
tmdb_data.info()
tmdb_data.head()

# calculate max string lengths for object columns
id_len = tmdb_data['imdb_id'].fillna('').map(len).max()
cert_len = tmdb_data['certification'].fillna('').map(len).max()
# create schema dictionary using SQLAlchemy datatype objects
tmdb_schema = {
    "imdb_id":String(id_len+1),
    "revenue":Float(),
    "budget":Float(),
    "numVotes":Text(cert_len+1)}

# add table to MySQL Server
tmdb_data.to_sql('tmdb_data',engine,dtype=tmdb_schema,
                   if_exists='replace',index=False)

# add primary key to title_basics table
engine.execute('ALTER TABLE tmdb_data ADD PRIMARY KEY (`imdb_id`);')

# execute query to display first 5 rows of table
q = """
SELECT * FROM tmdb_data
LIMIT 5;
"""
pd.read_sql(q,engine)

## Check Tables in the New MySQL Server

# execute query to display tables in the movies server
q = """
SHOW TABLES;
"""
pd.read_sql(q,engine)
